Fixed grammatical error, double 'of' line 46
The sentence originally read "A good distinction is that faceting gives you an overview description of all of the data that is currently..." it now reads "A good distinction is that faceting gives you an overview description of all the data that is currently..." by removal of the second "of" in the sentence. This is more concise language. 

* Data Carpentries Instructor Checkout Contribution *